+++ packages/brick_graphql_generators/lib/src/graphql_serialize.dart
@@ -23,9 +23,7 @@ class GraphqlSerialize extends GraphqlSerdesGenerator with JsonSerialize<Graphql
       final checker = checkerForType(field.type);
       final remoteName = providerNameForField(annotation.name, checker: checker);
       final columnInsertionType = _finalTypeForField(field.type);
-      final subfields = _subfieldsForType(field.type).fold<String>('{', (acc, subfield) {
-        return acc + "'subfield',";
-      });
+      final subfields = _subfieldsForType(field.type).map((f) => "'$f'").join(',');
 
       // T0D0 support List<Future<Sibling>> for 'association'
       fieldsToColumns.add('''
@@ -33,7 +31,7 @@ class GraphqlSerialize extends GraphqlSerdesGenerator with JsonSerialize<Graphql
             association: ${checker.isSibling || (checker.isIterable && checker.isArgTypeASibling)},
             documentNodeName: '$remoteName',
             iterable: ${checker.isIterable},
-            subfields: $subfields},
+            subfields: <String>{$subfields},
             type: $columnInsertionType,
           )''');
     }
@@ -65,12 +63,12 @@ class GraphqlSerialize extends GraphqlSerdesGenerator with JsonSerialize<Graphql
       return _subfieldsForType(checker.argType);
     }
 
-    if (checker.toJsonMethod != null) {
+    if (checker.toJsonMethod != null && checker.toJsonMethod!.returnType.isDartCoreMap) {
       if (type.element is ClassElement) {
         final klass = type.element as ClassElement;
         final subfields = klass.fields.where((field) {
           return field.isPublic &&
-              (field.isFinal || field.isConst || field.getter != null) &&
+              ((field.isFinal || field.isConst) && field.getter != null) &&
               !field.isStatic &&
               !field.type.isDartCoreFunction;
         }).where((field) => !FieldsForClass.isComputedGetter(field));