+++ packages/brick_graphql_generators/test/graphql_model_serdes_generator_test.dart
@@ -80,7 +80,5 @@ Future<void> generateAdapterExpectation(String filename, String output) async {
     annotation.annotation,
     null,
   );
-  print(generated);
-
   expect(generated.trim(), output.trim());
 }